One file per plugin. Fixes #657
