+++ lib/puppet-lint/plugins/check_comments/star_comments.rb
@@ -1,26 +1,3 @@
-# Public: Check the manifest tokens for any comments started with slashes
-# (//) and record a warning for each instance found.
-#
-# https://docs.puppet.com/guides/style_guide.html#comments
-PuppetLint.new_check(:slash_comments) do
-  def check
-    tokens.select { |token|
-      token.type == :SLASH_COMMENT
-    }.each do |token|
-      notify :warning, {
-        :message => '// comment found',
-        :line    => token.line,
-        :column  => token.column,
-        :token   => token,
-      }
-    end
-  end
-
-  def fix(problem)
-    problem[:token].type = :COMMENT
-  end
-end
-
 # Public: Check the manifest tokens for any comments encapsulated with
 # slash-asterisks (/* */) and record a warning for each instance found.
 #
@@ -41,7 +18,7 @@ PuppetLint.new_check(:star_comments) do
 
   def fix(problem)
     comment_lines = problem[:token].value.strip.split("\n").map(&:strip)
-    
+
     first_line = comment_lines.shift
     problem[:token].type = :COMMENT
     problem[:token].value = " #{first_line}"
