+++ lib/puppet-lint/plugins/check_documentation/documentation.rb
