+++ lib/puppet-lint/plugins/check_nodes/unquoted_node_name.rb
