+++ lib/puppet-lint/plugins/check_whitespace/arrow_alignment.rb
@@ -1,121 +1,3 @@
-# Public: Check the raw manifest string for lines containing hard tab
-# characters and record an error for each instance found.
-#
-# https://docs.puppet.com/guides/style_guide.html#spacing-indentation-and-whitespace
-PuppetLint.new_check(:hard_tabs) do
-  WHITESPACE_TYPES = Set[:INDENT, :WHITESPACE]
-
-  def check
-    tokens.select { |r|
-      WHITESPACE_TYPES.include?(r.type) && r.value.include?("\t")
-    }.each do |token|
-      notify :error, {
-        :message => 'tab character found',
-        :line    => token.line,
-        :column  => token.column,
-        :token   => token,
-      }
-    end
-  end
-
-  def fix(problem)
-    problem[:token].value.gsub!("\t", '  ')
-  end
-end
-
-# Public: Check the manifest tokens for lines ending with whitespace and record
-# an error for each instance found.
-#
-# https://docs.puppet.com/guides/style_guide.html#spacing-indentation-and-whitespace
-PuppetLint.new_check(:trailing_whitespace) do
-  def check
-    tokens.select { |token|
-      [:WHITESPACE, :INDENT].include?(token.type)
-    }.select { |token|
-      token.next_token.nil? || token.next_token.type == :NEWLINE
-    }.each do |token|
-      notify :error, {
-        :message => 'trailing whitespace found',
-        :line    => token.line,
-        :column  => token.column,
-        :token   => token,
-      }
-    end
-  end
-
-  def fix(problem)
-    prev_token = problem[:token].prev_token
-    next_token = problem[:token].next_token
-    prev_token.next_token = next_token
-    next_token.prev_token = prev_token unless next_token.nil?
-    tokens.delete(problem[:token])
-  end
-end
-
-# Public: Test the raw manifest string for lines containing more than 140
-# characters and record a warning for each instance found.  The only exceptions
-# to this rule are lines containing URLs and template() calls which would hurt
-# readability if split.
-#
-# https://docs.puppet.com/guides/style_guide.html#spacing-indentation-and-whitespace
-PuppetLint.new_check(:'140chars') do
-  def check
-    manifest_lines.each_with_index do |line, idx|
-      unless line =~ /:\/\// || line =~ /template\(/
-        if line.scan(/./mu).size > 140
-          notify :warning, {
-            :message => 'line has more than 140 characters',
-            :line    => idx + 1,
-            :column  => 140,
-          }
-        end
-      end
-    end
-  end
-end
-
-# Public: Test the raw manifest string for lines containing more than 80
-# characters. This is DISABLED by default and behaves like the default
-# 140chars check by excepting URLs and template() calls.
-#
-# https://docs.puppet.com/guides/style_guide.html#spacing-indentation-and-whitespace (older version)
-PuppetLint.new_check(:'80chars') do
-  def check
-    manifest_lines.each_with_index do |line, idx|
-      unless line =~ /:\/\// || line =~ /template\(/
-        if line.scan(/./mu).size > 80
-          notify :warning, {
-            :message => 'line has more than 80 characters',
-            :line    => idx + 1,
-            :column  => 80,
-          }
-        end
-      end
-    end
-  end
-end
-PuppetLint.configuration.send("disable_80chars")
-
-# Public: Check the manifest tokens for any indentation not using 2 space soft
-# tabs and record an error for each instance found.
-#
-# https://docs.puppet.com/guides/style_guide.html#spacing-indentation-and-whitespace
-PuppetLint.new_check(:'2sp_soft_tabs') do
-  def check
-    tokens.select { |r|
-      r.type == :INDENT
-    }.reject { |r|
-      r.value.length % 2 == 0
-    }.each do |token|
-      notify :error, {
-        :message => 'two-space soft tabs not used',
-        :line    => token.line,
-        :column  => token.column,
-      }
-    end
-  end
-end
-
 # Public: Check the manifest tokens for any arrows (=>) in a grouping ({}) that
 # are not aligned with other arrows in that grouping.
 #
